Plant Disease/requirements.txt
streamlit==1.16.0
tensorflow==2.13.0
pillow==8.4.0
numpy==1.23.5
